Update _hook.html.erb to prevent JS errors
When adding Paloma to a pre-existing codebase, users may encounter a problem where some pages (i.e. in Rails Engines) do not load the Paloma Javascript [i.e. no //= require paloma]  

The hook expects Paloma to be defined.  If Paloma is not defined there will be an error and JS execution stops.

This patch prevents the hook from causing the JS error.